Move to setup lib instead of App component
