+++ src/App.tsx
@@ -2,7 +2,6 @@ import {PortalProvider} from '@gorhom/portal';
 import React from 'react';
 import {LogBox} from 'react-native';
 import {GestureHandlerRootView} from 'react-native-gesture-handler';
-import Onyx from 'react-native-onyx';
 import {PickerStateProvider} from 'react-native-picker-select';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
 import '../wdyr';
@@ -31,8 +30,6 @@ import {WindowDimensionsProvider} from './components/withWindowDimensions';
 import Expensify from './Expensify';
 import useDefaultDragAndDrop from './hooks/useDefaultDragAndDrop';
 import OnyxUpdateManager from './libs/actions/OnyxUpdateManager';
-import * as Session from './libs/actions/Session';
-import * as Environment from './libs/Environment/Environment';
 import InitialUrlContext from './libs/InitialUrlContext';
 import {ReportAttachmentsProvider} from './pages/home/report/ReportAttachmentsContext';
 import type {Route} from './ROUTES';
@@ -41,29 +38,6 @@ type AppProps = {
     /** If we have an authToken this is true */
     url?: Route;
 };
-
-// For easier debugging and development, when we are in web we expose Onyx to the window, so you can more easily set data into Onyx
-if (window && Environment.isDevelopment()) {
-    window.Onyx = Onyx;
-
-    // We intentionally do not offer an Onyx.get API because we believe it will lead to code patterns we don't want to use in this repo, but we can offer a workaround for the sake of debugging
-    // @ts-expect-error TS233 - injecting additional utility for use in runtime debugging, should not be used in any compiled code
-    window.Onyx.get = function (key) {
-        return new Promise((resolve) => {
-            // eslint-disable-next-line rulesdir/prefer-onyx-connect-in-libs
-            const connectionID = Onyx.connect({
-                key,
-                callback: (value) => {
-                    Onyx.disconnect(connectionID);
-                    resolve(value);
-                },
-                waitForCollectionCallback: true,
-            });
-        });
-    };
-
-    window.setSupportToken = Session.setSupportAuthToken;
-}
 
 LogBox.ignoreLogs([
     // Basically it means that if the app goes in the background and back to foreground on Android,
